Fully integrate QtSerialBus into the UI
Allow selecting any CAN bus plugin provided by QtSerialBus,
not only SocketCAN, by appending them to the type combo box
after GVRET and KVASER.

For easier handling, get rid of the enum type and store the
connection as string. This allows similar handling for GVRET
and KVASER as well as for the QtSerialBus plugins.